WIP: Add multi-nesting and no-nesting examples for overseas futures and stocks
- Implemented multi-nesting strategy for overseas futures with three layers of conditions (all → any → at_least).
- Created multi-nesting strategy for overseas stocks using different logical operators (all, at_least, weighted).
- Developed no-nesting strategy for overseas futures combining multiple conditions at the same level.
- Added no-nesting strategy for overseas stocks with flat structure and all conditions evaluated together.
- Introduced single-nesting strategies for both overseas futures and stocks, allowing for a single level of nested conditions.
- Each example includes detailed descriptions, system settings, securities configurations, and order definitions.

+++ src/core/programgarden_core/bases/strategy.py
@@ -51,7 +51,7 @@ class BaseStrategyConditionResponseCommon(TypedDict):
     """EN: Additional payload produced by the condition implementation.
     KO: 조건 구현이 제공하는 추가 데이터입니다."""
 
-    weight: NotRequired[int]
+    weight: NotRequired[float]
     """EN: Optional weighting factor between 0 and 1 (defaults to 0).
     KO: 0과 1 사이에서 선택적으로 설정하는 가중치입니다 (기본값 0)."""
